Movement activity 24h
A security camera at the bank records movements during the day using a sequence in which each
member represents one hour. If there were any recorded movements in an hour,
the value of that clock in the sequence is 1, otherwise 0.
Write a function that receives such a string and prints messages according to the following criteria:
0 - 8 hours, print: “Normal activity”
9 am - 12 pm, print: “Increased activity”
12 and more hours, print: "Suspicious activity"