Vowel check and square number
Write a program that loads one positive integer and one letter from the user.
Then call the function which, depending on whether the letter consonant is entered or
vowel, calls another function: if the letter is a vowel, the function is called
which prints the square of the number received and does not change the number itself, and if it is a letter
consonant, a function is called that prints the square of the number received and changes itself
number (hence the change must be visible in the main program).